Mit maven bauen, java-packages eingeführt, Status der Buttons richtig setzen.
