--- AutoDrive Course Editor/src/MapNode.java
@@ -1,24 +1,0 @@
-import java.util.LinkedList;
-import java.util.TreeMap;
-
-public class MapNode {
-
-    public LinkedList<MapNode> incoming;
-    public LinkedList<MapNode> outgoing;
-    public TreeMap<MapMarker, MapNode> directions;
-    public double x, y, z;
-    public int id;
-
-    public MapNode(int id, double x, double y, double z) {
-        this.id = id;
-        this.x = x;
-        this.y = y;
-        this.z = z;
-        this.incoming = new LinkedList<>();
-        this.outgoing = new LinkedList<>();
-        this.directions = new TreeMap<>();
-    }
-
-
-
-}
