+++ AutoDrive Course Editor/src/main/java/de/adEditor/ADUtils.java
@@ -1,0 +1,18 @@
+package de.adEditor;
+
+import java.awt.geom.Rectangle2D;
+
+public class ADUtils {
+
+    public static Rectangle2D getNormalizedRectangleFor(double x, double y, double width, double height) {
+        if (width < 0) {
+            x = x + width;
+            width = -width;
+        }
+        if (height < 0) {
+            y = y + height;
+            height = -height;
+        }
+        return new Rectangle2D.Double(x,y,width,height);
+    }
+}
